refactor(core): split Auth Tab result/state types into their own files
Extracts LaunchAuthTabResult, ActivityState, PendingRequest,
AuthTabResult, AuthenticateUserResultContract, and
BrowserSwitchLaunchMode into their own files, one type per file. The
previously-private nested types (ActivityState, PendingRequest,
AuthTabResult, AuthenticateUserResultContract) become internal so they
can live at file scope while staying out of the public API;
LaunchAuthTabResult and BrowserSwitchLaunchMode keep the visibility and
@RestrictTo annotation they already had.

Also reverts two files touched by an earlier cleanup pass that
shouldn't have been in scope for the Auth Tab work, since both predate
this branch:

- BrowserSwitchPendingState.fromBase64() is back to its original
  runCatching-based implementation; the test added to cover the
  since-reverted rewrite is removed with it.
- PayPalClient.start()/vault() are back to their original single
  try/catch around session-fetch, checkout/vault launch, and the
  merchant callback together, rather than narrowing the catch to only
  the session-fetch step.

No functional change to the Auth Tab implementation itself; existing
unit tests and detekt continue to pass.


git-sptl-ai-lines: 175
git-sptl-human-lines: 0
git-sptl-unknown-lines: 0
git-sptl-ai-ratio: 1.00
git-sptl-ai-accepted: 175
git-sptl-ai-overridden: 0
git-sptl-tools: claude,codex:gpt-5.6-sol
git-sptl-signature: d21d4e778baedbbb0b9bd0c9e6a6ed559a9b4a665d9ba73742f7db950df15277
