+++ CorePayments/src/main/java/com/paypal/android/corepayments/browserswitch/AuthTabClient.kt
@@ -112,15 +112,3 @@ class AuthTabClient internal constructor(
         }
     }
 }
-
-@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-sealed class LaunchAuthTabResult {
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    data object Success : LaunchAuthTabResult()
-
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    data object ActivityNotFound : LaunchAuthTabResult()
-
-    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-    data class Failure(val error: Exception) : LaunchAuthTabResult()
-}
