+++ CorePayments/src/main/java/com/paypal/android/corepayments/browserswitch/AuthTabResult.kt
@@ -1,0 +1,8 @@
+package com.paypal.android.corepayments.browserswitch
+
+import android.net.Uri
+
+internal data class AuthTabResult(
+    val resultCode: Int,
+    val resultUri: Uri?,
+)
